A barebones navigation system & unit query.
You now get a list of units on the main "page", and can click on a unit to go
to the detail "page" for that unit. Routing currently very primitive.

The unit detail page doesn't display anything yet. I would like to reuse the
type definitions from server/units.ts but I'm not sure if that's possible when
they're covered in type-graphql annotations, so I need to look into that. Maybe
it's simpler to just leave the GQL query results untyped.